Bump mriqc version to 24.0.2
We add --no-datalad-get for mriqc images, which was not an option for 0.16

There was a note in the README that newer images could not run on CI,
but I found no job that ran the script in CI, so better to be correct.

I hand tested the script locally, and 24.0.2 worked as expected.